refactor 2 example tests and implement page object for HomePage.js
